fix(shard): make composite it's own folder
